np.projecteulerproblem11
